requirements.txt
# Interfaz gráfica
# tkinter viene incluido con Python

# Procesamiento de imágenes y OCR
Pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
numpy==1.26.3

# Base de datos
mysql-connector-python==8.3.0
python-dotenv==1.0.1

# Automatización web con Playwright
playwright==1.41.0

# Nota: Después de instalar, ejecutar:
# playwright install chromium
